(0721) refactor: 관리자 전체 신고조회 - 응답 객체 및 DTO 분리 기반의 계층간 책임 분리 리팩토링(0028)
